Create an editor class for the whole program This class will store the image (a collection of 'pixels' as a 2D array/matrix) Create an input class that can take a user's input for example I 5 5, and separate out the command (i) and the arguments (5 5). It will use this to map to a service object / command and pass int he arguments for processing. At the moment i'm storing executed commands into an array in case we want to add an undo function to them later then this sets us up for it, fully applying command design pattern. Until then then the commands will just be service objects.
The image class has a custom setter for pixels so it can reset it's dimenions instance variables to make sure its state stays up to date when modifying externally.

In context of the editor object - i'm passing in self to the commands so they can access the same image in memory off of the editor.