Create a clear image command.
This clear image command just needs to reset the image instance var on the editor to be a new image with the same dimensions, the way we can do it this way is because the image object defaults the pixels to O already for us. We just need to make sure the size is the same.