Create a fill region command
The fill region command will change the colour of a pixel and all adjascent matching ones (and the ones matching them and so on and so forth).

Use a scan line fill mechanism. Seems to be the most popular and sensible choice for such a use case. Recursion can cause errors with memory when reading in large images as you can expect.

Create a custom stack object so we can overwrite the push method, this lets us set a custom size for the array/stack and then cyle the pointer back.