Create a colour pixel command
This command changes the colour (c) of the selected pixel.

It does this by just overwriting the selected element in the array to be that of the user input.
Remember coords must be translated to indexes by minusing 1 since indexes start at 0 in ruby but user input will typically start at 1 when referring to coordinates