Create the draw vertical and draw horizontal commands.
These are similar with slight modifications to the condition in the execute method.
These 2 commands take 2 inputs for the range of how long they should, given a start and end and then (depending on which command you use) the location on the X or Y axis it should be on, then change the colour to be that the user input.

For these im just iterating through all the pixels and if it lies within the boundary then change it (inclusive). Not the best approach but for time's sake it will do for now.